fix(infra): resolve startup failures and type mismatches
- TenantRepository: fix ID type Integer → Long to match TenantDataModel
- SequentialIDGenerator: add @Lazy to break circular dependency
- AESGCMEncryptionService: add @Autowired for constructor disambiguation
- HibernateTenantFilter: wrap tenantId access in Optional.ifPresent()
- tenant-management-module.yaml: tenantId int32 → int64

+++ tenant-management/src/main/java/com/akademiaplus/interfaceadapters/TenantRepository.java
@@ -4,5 +4,5 @@ import com.akademiaplus.tenancy.TenantDataModel;
 import lombok.NonNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TenantRepository extends JpaRepository<@NonNull TenantDataModel, @NonNull Integer> {
+public interface TenantRepository extends JpaRepository<@NonNull TenantDataModel, @NonNull Long> {
 }
